Juxtaposition Assignment - first steps * setup juxtaposition assignment type * correct juxtaposition spelling 17 billion times. (#$%^) * refactor SelectionAssignmentEditView & SelectionAssignmentView functionality into generic views * add Juxtaposition views & url routing
