+++ media/js/app/projects/juxtaposition_assignment_edit.js
@@ -1,0 +1,10 @@
+/* global _: true, Backbone: true, CitationView: true */
+/* global showMessage: true, tinymce: true, tinymceSettings: true */
+// jscs:disable requireCamelCaseOrUpperCaseIdentifiers
+
+(function(jQuery) {
+    var global = this;
+
+    global.JuxtapositionAssignmentEditView = Backbone.View.extend({
+    });
+}(jQuery));
