+++ media/js/app/projects/juxtaposition_assignment_view.js
@@ -1,0 +1,19 @@
+/* global _: true, Backbone: true, djangosherd: true */
+/* global annotationList: true, CitationView: true */
+/* global Mustache: true, MediaThread: true, showMessage: true */
+// jscs:disable requireCamelCaseOrUpperCaseIdentifiers
+
+/**
+ * Listens For:
+ * Nothing
+ *
+ * Signals:
+ * Nothing
+ */
+
+(function(jQuery) {
+    var global = this;
+
+    global.JuxtapositionAssignmentView = Backbone.View.extend({
+    });
+}(jQuery));
